auth-server: server: handle_external_match: create gas_sponsorship module
